requirements.txt
playwright>=1.40.0
google-genai>=1.0.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
schedule>=1.2.0
httpx>=0.27.0
flask>=3.0.0
python-dotenv>=1.0.0
